feat: 좋아요 기능에 대한 테스크 코드 수행

+++ src/main/java/com/great/deploy/dolpin/domain/Comment.java
@@ -29,7 +29,7 @@ public class Comment {
     private Pins pins;
 
     private String contents;
-    private Long accountId;
+    private Integer accountId;
     private String nickName;
     private int recommendCount;
 
@@ -51,7 +51,7 @@ public class Comment {
     @UpdateTimestamp
     private LocalDateTime updateAt;
 
-    public Comment(Pins pins, String contents, Long accountId, String nickName) {
+    public Comment(Pins pins, String contents, Integer accountId, String nickName) {
         this.pins = pins;
         this.contents = contents;
         this.accountId = accountId;
